Implemented the sqlite database into order and order details page
havent done anything with the deliveryinfo page yet.
Will do it tomorrow

Help me review the structure of XERO.API.FillItems()
not sure if it could be better or not

+++ DeliveryAppWhiterocks/DeliveryAppWhiterocks/Models/Database/SQLite/InvoiceSQLite.cs
@@ -12,6 +12,6 @@ namespace DeliveryAppWhiterocks.Models.Database.SQLite
         public string InvoiceNumber { get; set; }
         public bool CompletedDeliveryStatus { get; set; }
         public string ContactID { get; set; }
-        public int Subtotal { get; set; }
+        public double Subtotal { get; set; }
     }
 }
